>[gh]: Relocate Pulumi POC's cluster-vault and npm workspace into its dir
The catalog/pulumi/cluster-vault/ component moved to
docs/experiments/20260702-pulumi-crossplane-evaluation/catalog/cluster-vault/,
since catalog/ is going to change and this component only exists to
support the POC. The npm workspace root (package.json,
package-lock.json) moved from the repo root into the same POC
directory so POC-only dependencies stop showing up in the repo-root
package.json.
