+++ docs/experiments/20260702-pulumi-crossplane-evaluation/catalog/cluster-vault/index.ts
