+++ docs/experiments/20260702-pulumi-crossplane-evaluation/package.json
@@ -1,0 +1,8 @@
+{
+	"name": "pulumi-crossplane-poc",
+	"private": true,
+	"workspaces": [
+		"catalog/cluster-vault",
+		"stack"
+	]
+}
